Updated codebase with context file, conventionally renamed providers, and altered export methods. Secondly, added an AuthProvider context file
